muted_users_ui: Move functions to more appropriate modules.
