+++ web/src/muted_users_ui.js
@@ -1,46 +1,11 @@
-import render_confirm_mute_user from "../templates/confirm_dialog/confirm_mute_user.hbs";
-
 import * as activity from "./activity";
-import * as channel from "./channel";
-import * as confirm_dialog from "./confirm_dialog";
-import {$t_html} from "./i18n";
 import * as message_lists from "./message_lists";
 import * as muted_users from "./muted_users";
 import * as overlays from "./overlays";
-import * as people from "./people";
 import * as pm_list from "./pm_list";
 import * as popovers from "./popovers";
 import * as recent_view_ui from "./recent_view_ui";
 import * as settings_muted_users from "./settings_muted_users";
-
-export function mute_user(user_id) {
-    channel.post({
-        url: "/json/users/me/muted_users/" + user_id,
-    });
-}
-
-export function confirm_mute_user(user_id) {
-    function on_click() {
-        mute_user(user_id);
-    }
-
-    const html_body = render_confirm_mute_user({
-        user_name: people.get_full_name(user_id),
-    });
-
-    confirm_dialog.launch({
-        html_heading: $t_html({defaultMessage: "Mute user"}),
-        help_link: "/help/mute-a-user",
-        html_body,
-        on_click,
-    });
-}
-
-export function unmute_user(user_id) {
-    channel.del({
-        url: "/json/users/me/muted_users/" + user_id,
-    });
-}
 
 export function rerender_for_muted_user() {
     message_lists.current.update_muting_and_rerender();
