+++ web/src/settings_muted_users.js
@@ -4,7 +4,6 @@ import render_muted_user_ui_row from "../templates/muted_user_ui_row.hbs";
 
 import * as ListWidget from "./list_widget";
 import * as muted_users from "./muted_users";
-import * as muted_users_ui from "./muted_users_ui";
 import * as people from "./people";
 import * as scroll_util from "./scroll_util";
 
@@ -52,7 +51,7 @@ export function set_up() {
         const user_id = Number.parseInt($row.attr("data-user-id"), 10);
 
         e.stopPropagation();
-        muted_users_ui.unmute_user(user_id);
+        muted_users.unmute_user(user_id);
     });
 
     populate_list();
